feat: started work on updating app to use mui

+++ src/components/entries/ViewEntryDetails.js
@@ -4,6 +4,8 @@ import { addEntryCategory, deleteEntry, getSingleEntry, removeEntryCategory } fr
 import { deleteComment, getCommentByEntryId } from "../../api/commentManager"
 import { useNavigate } from "react-router-dom"
 import { getAllCategories, getCategoriesByEntryId } from "../../api/categoryManager"
+import { Button, Dialog, DialogActions, DialogContent, DialogContentText, DialogTitle } from "@mui/material"
+import HighlightOffOutlinedIcon from '@mui/icons-material/HighlightOffOutlined'
 
 export const EntryDetails = ({ token }) => {
     const { entryId } = useParams()
@@ -13,13 +15,27 @@ export const EntryDetails = ({ token }) => {
     const [categories, setCategories] = useState([])
     const [selectedCategory, setSelectedCategory] = useState([])
     const [entryCategories, setEntryCategories] = useState([])
+    const [deleteConfirmationModalOpen, setDeleteConfirmationModalOpen] = useState(false)
 
-    const handleDelete = () => { // TODO add mui confirmation modal
-        deleteEntry(entryId, token)
-            .then(() => {
-                navigate("/entries")
-            })
+    const handleDeleteConfirmationModalOpen = () => {
+        setDeleteConfirmationModalOpen(true)
     }
+
+    const handleDelete = () => {
+        handleDeleteConfirmationModalOpen()
+
+    }
+
+    const handleDeleteConfirmationModalClose = (shouldDelete) => {
+        if (shouldDelete) {
+            deleteEntry(entryId, token)
+                .then(() => {
+                    navigate("/entries")
+                })
+        }
+        setDeleteConfirmationModalOpen(false)
+    }
+
 
     useEffect(() => {
         getSingleEntry(entryId, token)
@@ -76,10 +92,10 @@ export const EntryDetails = ({ token }) => {
             <h3 className="entry__title">{entry.title}</h3>
             <div className="entry__content">{entry.content}</div>
             <div className="entry__date">{entry.publication_date}</div>
-            <button className="btn btn-primary" onClick={() => {
+            <Button className="btn btn-primary" onClick={() => {
                 navigate(`/entries/edit/${entryId}`)
-            }}>Edit</button>
-            <button className="btn btn-primary" onClick={handleDelete}>Delete</button>
+            }}>Edit</Button>
+            <Button className="btn btn-primary" onClick={handleDelete}>Delete</Button>
         </section>
         <section className="categories">
             <h3>Current Categories</h3>
@@ -89,7 +105,7 @@ export const EntryDetails = ({ token }) => {
                     <div className="category" key={category.id}>
                         <div className="category__label">{category.label}</div>
                     </div>
-                    <button className="btn btn-primary" id={category.id} onClick={handleDeleteCategory}>X</button>
+                    <HighlightOffOutlinedIcon id={category.id} onClick={handleDeleteCategory} />
                     </>
                 })
             }
@@ -102,13 +118,13 @@ export const EntryDetails = ({ token }) => {
                         <div className="comment__author">{comment.author.user.username}</div>
                         <div className="comment__title"><h3>{comment.title}</h3></div>
                         <div className="comment__content">{comment.content}</div>
-                        <button className="btn btn-primary" id={comment.id} onClick={handleDeleteComment}>Delete</button>
+                        <Button className="btn btn-primary" id={comment.id} onClick={handleDeleteComment}>Delete</Button>
                     </div>
                 })
             }
-            <button className="btn btn-primary" onClick={() => {
+            <Button className="btn btn-primary" onClick={() => {
                 navigate(`/comments/create/${entryId}`)
-            }}>Add Comment</button>
+            }}>Add Comment</Button>
         </section>
         <section className="categories">
             <h3>Categories</h3>
@@ -123,8 +139,29 @@ export const EntryDetails = ({ token }) => {
                 }
             </select>
 
-            <button className="btn btn-primary" onClick={handleAddCategory}>Update Category</button>
+            <Button className="btn btn-primary" onClick={handleAddCategory}>Update Category</Button>
         </section>
+        <Dialog
+        open={deleteConfirmationModalOpen}
+        onClose={() => handleDeleteConfirmationModalClose(false)}
+        aria-labelledby="alert-dialog-title"
+        aria-describedby="alert-dialog-description"
+        >
+        <DialogTitle id="alert-dialog-title">
+            Are you sure you want to delete this entry?
+        </DialogTitle>
+        <DialogContent>
+            <DialogContentText id="alert-dialog-description">
+            This action cannot be undone.
+            </DialogContentText>
+        </DialogContent>
+        <DialogActions>
+            <Button onClick={() => handleDeleteConfirmationModalClose(false)}>Cancel</Button>
+            <Button onClick={() => handleDeleteConfirmationModalClose(true)} autoFocus>
+            Delete
+            </Button>
+        </DialogActions>
+        </Dialog>
         </>
     )
 }
